fixing

+++ kernel/trap.c
@@ -82,7 +82,7 @@ usertrap(void)
   else if (r_scause() == 15){ //only want write page fault
       pagetable_t page = p->pagetable;
       pte_t *pte = walk(page, r_stval(), 0);
-      if(cow_fault_handler(page, r_stval(), *pte) == -1){
+      if(cow_fault_handler(page, (uint64)r_stval(), *pte) == -1){
           printf("usertrap(): unexpected scause 0x%lx pid=%d\n", r_scause(), p->pid);
           printf("            sepc=0x%lx stval=0x%lx\n", r_sepc(), r_stval());
           setkilled(p);
